구조 수정
