+++ src/main/java/org/example/jjava_main/controller/UserController.java
@@ -3,8 +3,8 @@ package org.example.jjava_main.controller;
 import lombok.RequiredArgsConstructor;
 import org.example.jjava_main._core.util.Resp;
 import org.example.jjava_main.domain.user.User;
-import org.example.jjava_main.domain.user.UserRequest;
 import org.example.jjava_main.domain.user.UserService;
+import org.example.jjava_main.dto.UserRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
