+++ src/main/java/org/example/jjava_main/domain/user/UserRepository.java
@@ -3,19 +3,16 @@ package org.example.jjava_main.domain.user;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Repository
 public class UserRepository {
     private final EntityManager em;
 
-    @Transactional
-    public void levelUpdateById(UserLevel level, Integer id) {
-        String q = "UPDATE User u SET u.level = :level WHERE u.id = :id";
-        em.createQuery(q)
-                .setParameter("level", level)
-                .setParameter("id", id)
-                .executeUpdate();
+    public Optional<User> findById(Integer id) {
+        return Optional.ofNullable(em.find(User.class, id));
     }
+
 }
